% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimAlleles.R
\name{trimAlleles}
\alias{trimAlleles}
\title{Truncate allele names in haplotypes to the specified number of fields.}
\usage{
trimAlleles(haplotypes, reso)
}
\arguments{
\item{haplotypes}{Data frame of tilde-delimited haplotypes extracted from the famData provided to LDWrap()}

\item{reso}{An integer that specifies the number of fields to which colon-delimited allele names in famData should be truncated. The default value of 0 indicates no truncation. A value higher than the number of fields in the supplied allele data will result in no truncation.}
}
\value{
A data frame of two sets of tilde-delimited haplotypes.
}
\description{
This function accepts a dataframe of tilde-delimited haplotypes and trims colon-delimited names to the number of fields specified by 'reso'.
}
\note{
This function is for internal POULD use only.
}
\examples{
#
}
\keyword{trimAllele}
