% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cALD_FamilyFunctions.R
\name{writeVector}
\alias{writeVector}
\title{Exporting Haplotype Vectors}
\usage{
writeVector(nLoc1, Res, hapVec, numSamp, writeName, genPhase, Prefix, vDir)
}
\arguments{
\item{nLoc1}{Number of alleles at locus 1}

\item{Res}{Data frame consisting of Locus Number, Locus Name, Allele Name, Allele Count, Allele Frequency and Frequency^2}

\item{hapVec}{Haplotype vector}

\item{numSamp}{The dataset's 2N value}

\item{writeName}{The specified name for the vector}

\item{genPhase}{Boolean describing phased (TRUE) or unphased (FALSE) analysis}

\item{Prefix}{The specified prefix for the written vector file}

\item{vDir}{The directory into which the vector file should be written.}
}
\description{
This function writes the haplotype vector and any accessory information to a text file.
}
\note{
This function is for internal POULD use only.
}
\examples{
#
}
\keyword{writeVector}
