% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cALD_FamilyFunctions.R
\name{LDWrap}
\alias{LDWrap}
\title{Parser for CSV-formatted 17th IHIW Family HLA Haplotype Data}
\usage{
LDWrap(famData, threshold = 10, phased = TRUE,
  frameName = "hla-family-data")
}
\arguments{
\item{famData}{A data frame or CSV formatted file that contains the two columns named "Gl String" and "Relation". Other columns can be included (in any order), but will not impact the analysis. The Relation column can contain any data; however anything other than "Relation=child" will be included in the LD analyses. The Gl String column should consist of two tilde (~) delimited HLA haplotypes conneced by a plus (+) sign (GL String format). The name of this file will serve as the basis for the name of the LD result files.}

\item{threshold}{An integer that specifies the minimnum number of subjects allowed for the analysis of a locus-pair. The default value is 10. If the number of subjects with haplotypes for a locus pair is less than the threshold, the *_LD_results.csv file will contain 'Not Calculated' 'Subject Threshold=##' 'Complete subjects=#' '.' in columns 2-5 for that locus pair, where ## is the set threshold and # is the number of subjects. Column 6 will be empty.}

\item{phased}{A boolean that determines if the LD calculations should be performed for phased data (TRUE) or unphased data (FALSE). If phased=FALSE, the EM algorithm is used to estimate haplotypes for the data in the Gl String column.}

\item{frameName}{A descriptor for the data frame of family data provided. Defaults to "hla-family-data". This value is not used if a CSV file is provided.}
}
\description{
A wrapper for parsing 17th International HLA and Immunogenetics Workshop Family Data files containing phased two-locus classical HLA locus haplotype data recorded in GL String format. Extracts all pairs of loci from GL String formatted haplotypes, and passes paired-genotype data to the cALD function.
}
\details{
This function coerces cALD() to generate a haplotype vector file for each locus pair analyzed, and generates a single LD results file containing LD values for all locus pairs, along with the number of haplotypes tested, one locus pair per row. The LD results file will contain six columns ("Loc1~Loc2","D'","Wn","WLoc1/Loc2","WLoc2/Loc1","N_Haplotypes"), and will be named "<filename prefix>_<Phased/Unphased>_LD_results.csv".
}
\note{
This function attemtps to peform these LD calculations for all pairs of the classical HLA loci -- HLA-A, -C, -B, -DRB1, -DRB3, -DRB4, -DRB5, -DQA1, -DQB1, -DPA1 and -DPB1. If a haplotype dataset does not include all of these loci, the *LD_results.csv file will include rows for locus pairs for which no data was avialable.

When at least one locus in a locus pair is monomorphic, no LD calculations will be performed, and column 5 of the results for that locus pair will identify the monomorphic loci.
}
\examples{
# Analyze the included example data
LDWrap(hla.hap.demo,frameName="HLADemo") 
}
\keyword{ldwrap}
\keyword{ldwrapper}
\keyword{wrapper}
